+++ server/src/main/kotlin/net/horizonsend/ion/server/IonServer.kt
@@ -103,6 +103,7 @@ class IonServer : JavaPlugin() {
 
 					pluginManager.registerEvents(CityNPCs, this)
 
+					SpaceGenerationManager.primeGeneration()
 					commandManager.commandCompletions.registerCompletion("wreckSchematics") { context ->
 						SpaceGenerationManager.getGenerator(
 							(context.player.world as CraftWorld).handle
