+++ server/src/main/kotlin/net/horizonsend/ion/server/features/space/generation/generators/SpaceGenerator.kt
@@ -3,44 +3,32 @@ package net.horizonsend.ion.server.features.space.generation.generators
 import com.sk89q.jnbt.NBTInputStream
 import com.sk89q.worldedit.extent.clipboard.Clipboard
 import com.sk89q.worldedit.extent.clipboard.io.SpongeSchematicReader
-import com.sk89q.worldedit.math.BlockVector3
 import net.horizonsend.ion.server.IonServer
 import net.horizonsend.ion.server.configuration.ServerConfiguration
 import net.horizonsend.ion.server.features.space.encounters.Encounter
 import net.horizonsend.ion.server.features.space.encounters.Encounters
-import net.horizonsend.ion.server.features.space.generation.BlockSerialization
 import net.horizonsend.ion.server.miscellaneous.NamespacedKeys
 import net.horizonsend.ion.server.miscellaneous.WeightedRandomList
 import net.minecraft.core.BlockPos
 import net.minecraft.core.registries.Registries
 import net.minecraft.nbt.CompoundTag
-import net.minecraft.nbt.ListTag
 import net.minecraft.nbt.NbtIo
 import net.minecraft.nbt.NbtUtils
 import net.minecraft.server.level.ServerLevel
-import net.minecraft.world.level.ChunkPos
 import net.minecraft.world.level.block.Blocks
 import net.minecraft.world.level.block.state.BlockState
-import net.minecraft.world.level.levelgen.Heightmap
-import net.starlegacy.util.Tasks
 import net.starlegacy.util.nms
-import net.starlegacy.util.time
 import net.starlegacy.util.timing
-import net.starlegacy.util.toBukkitBlockData
-import net.starlegacy.util.worldEditSession
 import org.bukkit.Bukkit
 import org.bukkit.Chunk
 import org.bukkit.craftbukkit.v1_19_R2.CraftChunk
 import org.bukkit.persistence.PersistentDataType
 import org.bukkit.util.noise.SimplexOctaveGenerator
 import java.io.ByteArrayInputStream
-import java.io.ByteArrayOutputStream
 import java.io.File
 import java.io.FileInputStream
 import java.util.Random
-import java.util.concurrent.CompletableFuture
 import java.util.zip.GZIPInputStream
-import kotlin.math.abs
 import kotlin.math.floor
 import kotlin.math.pow
 import kotlin.math.sqrt
@@ -58,6 +46,12 @@ class SpaceGenerator(
 	val timing = timing("Space Generation")
 	val random = Random(serverLevel.seed)
 
+	abstract class SpaceGenerationData {
+		abstract val x: Int
+		abstract val y: Int
+		abstract val z: Int
+	}
+
 	/**
 	 * This class contains information passed to the generation function.
 	 * @param [x, y ,z] Origin of the asteroid.
@@ -66,16 +60,16 @@ class SpaceGenerator(
 	 * @param octaves The number of octaves of noise to apply. Generally 1, but higher for small asteroids. Increases roughness.
 	 **/
 	data class AsteroidGenerationData(
-		val x: Int,
-		val y: Int,
-		val z: Int,
+		override val x: Int,
+		override val y: Int,
+		override val z: Int,
 		val palette: WeightedRandomList<BlockState>,
 		val size: Double,
 		val octaves: Int
-	)
+	) : SpaceGenerationData()
 
 	// ASTEROIDS SECTION
-	private val oreMap: Map<String, BlockState> = configuration.ores.associate {
+	val oreMap: Map<String, BlockState> = configuration.ores.associate {
 		it.material to Bukkit.createBlockData(it.material).nms
 	}
 
@@ -85,273 +79,10 @@ class SpaceGenerator(
 	// Palettes weighted
 	private val weightedPalettes = configuration.paletteWeightedList()
 
-	// 	private val weightedPalettes = configuration.blockPalettes.associateWith { it.materialWeights() }
-	private val weightedOres = oreWeights()
+	val weightedOres = oreWeights()
 
 	// Multiple of the radius of the asteroid to mark chunks as might contain an asteroid
-	private val searchRadius = 1.25
-
-	fun generateAsteroid(
-		asteroid: AsteroidGenerationData
-	) {
-		timing.time {
-			Tasks.async {
-				// save some time
-				val sizeFactor = asteroid.size / 15
-				val shapingNoise = SimplexOctaveGenerator(random, 1)
-				val materialNoise = SimplexOctaveGenerator(random, 1)
-				materialNoise.setScale(0.15 / sqrt(sizeFactor))
-
-				val radiusSquared = asteroid.size * asteroid.size
-
-				// Get every chunk section covered by the asteroid.
-				val coveredChunks = mutableMapOf<ChunkPos, List<Int>>()
-
-				// generate ranges ahead of time
-				val xRange =
-					IntRange(asteroid.x - (asteroid.size * searchRadius).toInt(), asteroid.x + (asteroid.size * searchRadius).toInt())
-				val zRange =
-					IntRange(asteroid.z - (asteroid.size * searchRadius).toInt(), asteroid.z + (asteroid.size * searchRadius).toInt())
-				val yRange = IntRange(
-					(asteroid.y - (asteroid.size * searchRadius).toInt()).coerceAtLeast(serverLevel.minBuildHeight),
-					(asteroid.y + (asteroid.size * searchRadius).toInt()).coerceAtMost(serverLevel.maxBuildHeight)
-				)
-
-				val chunkXRange = IntRange(xRange.first.shr(4), xRange.last.shr(4))
-				val chunkZRange = IntRange(zRange.first.shr(4), zRange.last.shr(4))
-				val chunkYRange = IntRange(yRange.first.shr(4), yRange.last.shr(4) + serverLevel.minBuildHeight.shr(4))
-
-				for (chunkPosX in chunkXRange) {
-					val xSqr = (chunkPosX - asteroid.x.shr(4)) * (chunkPosX - asteroid.x.shr(4))
-
-					for (chunkPosZ in chunkZRange) {
-						val zSqr = (chunkPosZ - asteroid.z.shr(4)) * (chunkPosZ - asteroid.z.shr(4))
-						val circle = xSqr + zSqr
-
-						if (circle >= radiusSquared) continue // if out of equatorial radius continue
-
-						val sections = mutableListOf<Int>()
-
-						for (chunkSectionY in chunkYRange) {
-							val ySqr = (chunkSectionY - asteroid.y.shr(4)) * (chunkSectionY - asteroid.y.shr(4))
-
-							if ((circle + ySqr) <= radiusSquared) {
-								sections += chunkSectionY
-							}
-						}
-
-						coveredChunks[ChunkPos(chunkPosX, chunkPosZ)] = sections
-					}
-				}
-				// Covered chunks acquired
-
-				// For each covered chunk
-				for ((nmsChunkPos, sectionList) in coveredChunks) {
-					serverLevel.world.getChunkAtAsync(nmsChunkPos.x, nmsChunkPos.z)
-						.thenAcceptAsync completedChunk@{ bukkitChunk ->
-							val completableBlocksChanged: CompletableFuture<List<BlockPos>> = CompletableFuture()
-							val chunkBlocksChanged = mutableListOf<BlockPos>()
-
-							val levelChunk = (bukkitChunk as CraftChunk).handle
-							val newSections = mutableListOf<CompoundTag>()
-
-							val chunkMinX = levelChunk.pos.x.shl(4)
-							val chunkMinZ = levelChunk.pos.z.shl(4)
-
-							for (sectionPos in sectionList) {
-								val levelChunkSection = levelChunk.sections[sectionPos]
-
-								val palette = mutableSetOf<BlockState>()
-								val storedBlocks = arrayOfNulls<Int>(4096)
-								var index = 0
-								val sectionMinY = levelChunkSection.bottomBlockY()
-
-								palette.add(Blocks.AIR.defaultBlockState())
-								val paletteListTag = ListTag()
-
-								for (x in 0..15) {
-									val worldX = chunkMinX + x
-									val worldXDouble = worldX.toDouble()
-									val xSquared = (worldXDouble - asteroid.x) * (worldXDouble - asteroid.x)
-
-									for (z in 0..15) {
-										val worldZ = chunkMinZ + z
-										val worldZDouble = worldZ.toDouble()
-										val zSquared = (worldZDouble - asteroid.z) * (worldZDouble - asteroid.z)
-
-										for (y in 0..15) {
-											val worldY = sectionMinY + y
-											val worldYDouble = worldY.toDouble()
-											val ySquared = (worldYDouble - asteroid.y) * (worldYDouble - asteroid.y)
-
-											val existingBlock = levelChunkSection.getBlockState(x, y, z)
-
-											if (!existingBlock.isAir) {
-												palette.add(existingBlock)
-												storedBlocks[index] = palette.indexOf(existingBlock)
-												index++
-												continue
-											}
-
-											var block: BlockState? =
-												checkBlockPlacement(
-													worldXDouble,
-													worldYDouble,
-													worldZDouble,
-													xSquared,
-													ySquared,
-													zSquared,
-													asteroid,
-													sizeFactor,
-													shapingNoise,
-													materialNoise
-												)
-
-											if (
-												(
-													random.nextDouble(0.0, 1.0) <= configuration.oreRatio &&
-														block != null
-													) && !block.isAir
-											) {
-												val ore = weightedOres[random.nextInt(0, weightedOres.size - 1)]
-												block = oreMap[ore.material]
-											}
-
-											if (block != null) {
-												palette.add(block)
-												storedBlocks[index] = palette.indexOf(block)
-
-												levelChunkSection.setBlockState(
-													x,
-													y,
-													z,
-													block
-												)
-											} else {
-												storedBlocks[index] = 0
-											}
-
-											// needs to be run on main thread
-											chunkBlocksChanged += BlockPos(worldX, worldY, worldZ)
-
-											index++
-										}
-									}
-								}
-
-								if (storedBlocks.all { it == 0 }) continue // don't write it if it's all empty
-
-								palette.forEach { blockState -> paletteListTag.add(NbtUtils.writeBlockState(blockState)) }
-
-								val intArray = storedBlocks.requireNoNulls().toIntArray()
-								newSections += BlockSerialization.formatSection(
-									sectionPos,
-									intArray,
-									paletteListTag
-								)
-							}
-							// start broadcasting the chunk information
-							completableBlocksChanged.complete(chunkBlocksChanged)
-
-							// Chunk is empty, everything else is unnecessary
-							if (newSections.isEmpty()) return@completedChunk
-
-							// data serialization
-							val completableBlockData = CompletableFuture<ByteArray>()
-							val existingSerializedAsteroidData =
-								BlockSerialization.readChunkCompoundTag(bukkitChunk, NamespacedKeys.STORED_CHUNK_BLOCKS)
-
-							val formattedSections = existingSerializedAsteroidData.getList("sections", 10) // list of CompoundTag (10)
-							formattedSections.addAll(newSections)
-							val storedChunkBlocks =
-								BlockSerialization.formatChunk(formattedSections, spaceGenerationVersion)
-							val outputStream = ByteArrayOutputStream()
-							NbtIo.writeCompressed(storedChunkBlocks, outputStream)
-							completableBlockData.complete(outputStream.toByteArray())
-							// end data serialization
-
-							// updating chunk information
-							Heightmap.primeHeightmaps(levelChunk, Heightmap.Types.values().toSet())
-							levelChunk.isUnsaved = true
-
-							// needs to be synchronized or multiple asteroids in a chunk cause issues
-							completableBlockData.thenAccept { blockData ->
-								// Write PDC sync to avoid issues
-								Tasks.sync {
-									bukkitChunk.persistentDataContainer.set(
-										NamespacedKeys.STORED_CHUNK_BLOCKS,
-										PersistentDataType.BYTE_ARRAY,
-										blockData
-									)
-								}
-							}
-
-							completableBlocksChanged.thenAccept { completed ->
-								Tasks.sync {
-									// broadcast updates synchronously
-									for (blockPos in completed) {
-										levelChunk.playerChunk?.blockChanged(blockPos)
-									}
-
-									levelChunk.playerChunk?.broadcastChanges(levelChunk)
-								}
-							}
-							// end chunk update
-						}
-				}
-			}
-		}
-	}
-
-	/**
-	 * Places the asteroid block, if inside, and returns the block
-	 **/
-	private fun checkBlockPlacement(
-		worldX: Double,
-		worldY: Double,
-		worldZ: Double,
-		worldXSquared: Double,
-		worldYSquared: Double,
-		worldZSquared: Double,
-		asteroid: AsteroidGenerationData,
-		sizeFactor: Double,
-		shapingNoise: SimplexOctaveGenerator,
-		materialNoise: SimplexOctaveGenerator
-	): BlockState? {
-		// Calculate a noise pattern with a minimum at zero, and a max peak of the size of the materials list.
-		val paletteSample = (
-			(
-				materialNoise.noise(
-					worldX,
-					worldY,
-					worldZ,
-					1.0,
-					1.0,
-					true
-				) + 1
-				) / 2
-			)
-
-		// Full noise is used as the radius of the asteroid, and it is offset by the noise of each block pos.
-		var fullNoise = 0.0
-		val initialScale = 0.015 / sizeFactor.coerceAtLeast(1.0)
-
-		for (octave in 0..asteroid.octaves) {
-			shapingNoise.setScale(initialScale * (octave + 1.0).pow(2.25 + (sizeFactor / 2.25).coerceAtMost(0.5)))
-
-			val offset = abs(
-				shapingNoise.noise(worldX, worldY, worldZ, 0.0, 1.0, false)
-			) * (asteroid.size / (octave + 1.0).pow(2.25))
-
-			fullNoise += offset
-		}
-
-		fullNoise *= fullNoise
-		// Continue if block is not inside any asteroid
-		if (worldXSquared + worldYSquared + worldZSquared >= fullNoise) return null
-
-		return asteroid.palette.getEntry(paletteSample)
-	}
+	val searchRadius = 1.25
 
 	/**
 	 * Generates an asteroid with optional specification for the parameters
@@ -432,228 +163,18 @@ class SpaceGenerator(
 		return SpongeSchematicReader(NBTInputStream(GZIPInputStream(FileInputStream(file)))).read()
 	}
 
-	fun generateWreck(wreck: WreckGenerationData) {
-		val encounter = wreck.encounter?.getEncounter()
-		Tasks.async {
-			val schematic: Clipboard = schematicMap[wreck.schematicName]!!
-
-			serverLevel.world.worldEditSession(true) {
-				val region = schematic.region.clone()
-				val targetBlockVector = BlockVector3.at(wreck.x, wreck.y, wreck.z)
-				val offset = targetBlockVector.subtract(schematic.origin)
-
-				region.shift(offset)
-				val sectionsSet = mutableSetOf<Int>()
-
-				val wreckChest = CompletableFuture<BlockPos>()
-
-				for (
-				y in (region.boundingBox.minimumY - serverLevel.minBuildHeight)..(region.boundingBox.maximumY - serverLevel.minBuildHeight)
-				) {
-					sectionsSet.add(y.shr(4))
-				}
-
-				for (chunkPosWE in region.chunks) {
-					val chunkOriginX = chunkPosWE.x * 16
-					val chunkOriginZ = chunkPosWE.z * 16
-
-					serverLevel.world.getChunkAtAsync(chunkPosWE.x, chunkPosWE.z)
-						.thenAcceptAsync completedChunk@{ bukkitChunk ->
-							val completableBlocksChanged: CompletableFuture<List<BlockPos>> = CompletableFuture()
-							val chunkBlocksChanged = mutableListOf<BlockPos>()
-
-							val levelChunk = (bukkitChunk as CraftChunk).handle
-							val newSections = mutableListOf<CompoundTag>()
-
-							for (section in sectionsSet) {
-								val levelChunkSection = levelChunk.sections[section]
-								val bottomY = levelChunkSection.bottomBlockY()
-
-								val palette = mutableSetOf<BlockState>()
-								palette.add(Blocks.AIR.defaultBlockState())
-								val storedBlocks = arrayOfNulls<Int>(4096)
-
-								var index = 0
-
-								for (x in 0..15) {
-									val worldX = x + chunkOriginX
-
-									for (z in 0..15) {
-										val worldZ = z + chunkOriginZ
-
-										for (y in 0..15) {
-											val worldY = bottomY + y
-											val schematicRelative = BlockVector3.at(worldX, worldY, worldZ).subtract(offset)
-
-											val baseBlock = schematic.getFullBlock(schematicRelative)
-											var blockState: BlockState = baseBlock.toImmutableState().toBukkitBlockData().nms
-											val blockNBT = if (blockState.hasBlockEntity()) baseBlock.nbtData else null
-
-											val existingBlock = levelChunkSection.getBlockState(x, y, z)
-
-											if (!existingBlock.isAir) {
-												palette.add(existingBlock)
-												storedBlocks[index] = palette.indexOf(existingBlock)
-												index++
-												continue
-											}
-
-											if (blockState.isAir) {
-												storedBlocks[index] = 0
-												index++
-												continue
-											}
-
-											encounter?.let {
-												blockNBT?.let {
-													if (blockState.block == Blocks.CHEST &&
-														blockNBT.getString("CustomName").contains(
-																"Encounter Chest",
-																true
-															)
-													) {
-														wreckChest.complete(BlockPos(worldX, worldY, worldZ))
-														blockState = encounter.constructChestState()
-													}
-												}
-											}
-
-											// needs to be run on main thread
-											chunkBlocksChanged += BlockPos(worldX, worldY, worldZ)
-
-											palette.add(blockState)
-											storedBlocks[index] = palette.indexOf(blockState)
-											levelChunkSection.setBlockState(
-												x,
-												y,
-												z,
-												blockState
-											)
-
-											index++
-										}
-									}
-								}
-
-								if (storedBlocks.all { it == 0 }) continue
-								// don't write it if it's all empty
-
-								val paletteListTag = ListTag()
-								palette.mapTo(paletteListTag) { blockState -> NbtUtils.writeBlockState(blockState) }.toList()
-
-								// Add the stored blocks to the list of sections
-								val intArray = storedBlocks.requireNoNulls().toIntArray()
-								newSections += BlockSerialization.formatSection(
-									section,
-									intArray,
-									paletteListTag
-								)
-							}
-
-							// start broadcasting the chunk information
-							completableBlocksChanged.complete(chunkBlocksChanged)
-
-							// Chunk is empty, everything else is unnecessary
-							if (newSections.isEmpty()) return@completedChunk
-
-							val completableBlockData = CompletableFuture<ByteArray>()
-
-							// Check chunk for existing spacegen blocks
-							val existingChunkBlocks =
-								BlockSerialization.readChunkCompoundTag(
-									bukkitChunk,
-									NamespacedKeys.STORED_CHUNK_BLOCKS
-								)
-
-							// list of CompoundTag (10)
-							val formattedSections = existingChunkBlocks.getList("sections", 10)
-							formattedSections.addAll(newSections)
-
-							val blocksOutputStream = ByteArrayOutputStream()
-							NbtIo.writeCompressed(
-								BlockSerialization.formatChunk(formattedSections, spaceGenerationVersion),
-								blocksOutputStream
-							)
-
-							completableBlockData.complete(blocksOutputStream.toByteArray())
-							// End wreck block storage
-
-							// updating chunk information
-							Heightmap.primeHeightmaps(levelChunk, Heightmap.Types.values().toSet())
-							levelChunk.isUnsaved = true
-
-							// Write wreck data to chunk
-							wreckChest.thenAccept { chestPos ->
-								Tasks.sync {
-									val existingWrecksBaseTag = BlockSerialization
-										.readChunkCompoundTag(
-											levelChunk.bukkitChunk,
-											NamespacedKeys.WRECK_ENCOUNTER_DATA
-										)
-									val existingWrecks =
-										existingWrecksBaseTag.getList("wrecks", 10) // list of compound tags (10)
-
-									// Won't be null if the encounter is not null, unless someone messed up
-									wreck.encounter?.let {
-										existingWrecks += it.nms(
-											chestPos.x,
-											chestPos.y,
-											chestPos.z
-										)
-									}
-
-									val newFinishedData = CompoundTag()
-									newFinishedData.put("wrecks", existingWrecks)
-									val wreckDataOutputStream = ByteArrayOutputStream()
-									NbtIo.writeCompressed(newFinishedData, wreckDataOutputStream)
-
-									levelChunk.bukkitChunk.persistentDataContainer.set(
-										NamespacedKeys.WRECK_ENCOUNTER_DATA,
-										PersistentDataType.BYTE_ARRAY,
-										wreckDataOutputStream.toByteArray()
-									)
-								}
-							}
-
-							completableBlockData.thenAccept {
-								Tasks.sync {
-									levelChunk.bukkitChunk.persistentDataContainer.set(
-										NamespacedKeys.STORED_CHUNK_BLOCKS,
-										PersistentDataType.BYTE_ARRAY,
-										blocksOutputStream.toByteArray()
-									)
-								}
-							}
-
-							// needs to be synchronized or multiple asteroids in a chunk cause issues
-							completableBlocksChanged.thenAccept { completed ->
-								Tasks.sync {
-									// broadcast updates synchronously
-									for (blockPos in completed) {
-										levelChunk.playerChunk?.blockChanged(blockPos)
-									}
-
-									levelChunk.playerChunk?.broadcastChanges(levelChunk)
-								}
-							}
-						}
-				}
-			}
-		}
-	}
-
 	/**
 	 * This information is not serialized. It is used in the generation of the wreck.
 	 * @param schematicName The name of the schematic file referenced, not including file extension
 	 * @param encounter The optional encounter data.
 	 **/
 	data class WreckGenerationData(
-		val x: Int,
-		val y: Int,
-		val z: Int,
+		override val x: Int,
+		override val y: Int,
+		override val z: Int,
 		val schematicName: String,
 		val encounter: WreckEncounterData?
-	) {
+	) : SpaceGenerationData() {
 		/**
 		 * This is serialized and stored in the chunk alongside the wreck.
 		 *
